move some tests from chatbot to boost
