+++ packages/client/src/views/chatbot/boost.ts
@@ -1,4 +1,3 @@
-import { Context, Language } from "decorator-shared/params";
 import loadExternalScript from "../../helpers/load-external-script";
 import { env, param } from "../../params";
 
@@ -9,13 +8,25 @@ export const initBoost = (): Promise<{ show: () => void }> =>
             return { show: () => {} };
         }
 
-        const boost = window.boostInit(
-            env("BOOST_ENV"),
-            buildBoostConfig({
-                context: param("context"),
-                language: param("language"),
-            }),
-        );
+        const boost = window.boostInit(env("BOOST_ENV"), {
+            chatPanel: {
+                settings: {
+                    removeRememberedConversationOnChatPanelClose: true,
+                    openTextLinksInNewTab: true,
+                },
+                styling: { buttons: { multiline: true } },
+                header: {
+                    filters: {
+                        filterValues:
+                            param("context") === "arbeidsgiver"
+                                ? "arbeidsgiver"
+                                : param("language") === "nn"
+                                  ? "nynorsk"
+                                  : "bokmal",
+                    },
+                },
+            },
+        });
 
         boost.chatPanel.addEventListener("setFilterValue", (event) => {
             boost.chatPanel.setFilterValues(event.detail.filterValue);
@@ -26,32 +37,6 @@ export const initBoost = (): Promise<{ show: () => void }> =>
 
         return { show: boost.chatPanel.show };
     });
-
-const buildBoostConfig = ({
-    context,
-    language,
-}: {
-    context: Context;
-    language: Language;
-}) => ({
-    chatPanel: {
-        settings: {
-            removeRememberedConversationOnChatPanelClose: true,
-            openTextLinksInNewTab: true,
-        },
-        styling: { buttons: { multiline: true } },
-        header: {
-            filters: {
-                filterValues:
-                    context === "arbeidsgiver"
-                        ? "arbeidsgiver"
-                        : language === "nn"
-                          ? "nynorsk"
-                          : "bokmal",
-            },
-        },
-    },
-});
 
 export const hasActiveConversation = (): boolean =>
     null !==
